requirements.txt
asgiref==3.8.1
django==4.2.18
djangorestframework==3.15.2
django-cors-headers==4.6.0
dotenv-python==0.0.1
sqlparse==0.5.3
typing-extensions==4.12.2